requirements.txt
streamlit==1.45.1
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.4
trafilatura==2.0.0
newspaper3k==0.2.8 